separate memory_store example RPC server to multiple files by handlers groups
